+++ R/include_css_js.R
@@ -12,7 +12,7 @@ include_css_files <- function(package = "teal", pattern = "*") {
 
   css_files <- list.files(system.file("css", package = package, mustWork = TRUE), pattern = pattern, full.names = TRUE)
 
-  return(lapply(css_files, includeCSS))
+  return(singleton(lapply(css_files, includeCSS)))
 }
 
 #' Include `JS` files from `/inst/js/` package directory to application header